backend/requirements.txt
django~=5.0.3
django_restframework
djangorestframework~=3.15.1
django-cors-headers
psycopg2
Pillow